Chang header element to nav-bar
My last push didn't work. it seems like the header footer section is not being applied in the css. I renamed header to nav-bar to see if this makes a difference. If not, I'm going to ask for help in the Odin Discord channel.

update: that seems to have done the trick!